Set output_fname to none if no target grid is being written
The topo tool consists of two runs, the first is to map the cube sphere
topography to a cube sphere with smooth topography, the second is to map the
smooth topography cube sphere to the target grid.

The way the tool achieves this is by running cube_to_target twice from the
Makefile. The Makefile also generates the namelist for each run as well, then
depending on the namelist cube_to_target either generates the smooth topography
cube sphere or maps cube sphere to the target grid.

During the first run the namelist variable, output_grid, is not set, but
cube_to_sphere prints it, which prints garbage characters to the string.
Although this happens on occasion, sometimes these garbage characters can trash
the rest of the output of a terminal making it impossible to read and the user
must type 'reset' to return their terminal to normal.

This commit ensures that output_grid always is set, thus, does not print
garbage characters and does not output it during the first run. As well, it
also makes the output more readable as garbage characters can give the illusion
that something is not correctly working.