Fix bug that prevented read of smooth topo cube with lregional_refinement
Before this change the first run to generate the smooth topo cube sphere would
not add '_RR' to the name of the smooth topo cube file, but the second run,
which tried to read in that same file, would. Thus, the code would fail due to
different expected filenames.

This commit removes the '_RR' tag in proctag. This means that _RR will not
appear in the file.